move test to NumberFieldMapperTests
